requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
httpx>=0.27.0
python-multipart>=0.0.9
python-dotenv>=1.0.1
torch>=2.2.0
torchvision>=0.17.0
Pillow>=10.0.0
